// projects/common/src/lib/interfaces/status.interface.ts
export interface StatusInterface {
    id: string;
    entity_id: string;
    organization_id: string;
    account_id: string;
    request_id: string;
    status: string;
    message: string;
    starttime: string;
    endtime: string;
    created: string;
    updated: string;
    deleted: string;
}